+++ database/migrations/2019_01_23_000001_create_tables_table.php
@@ -14,7 +14,7 @@ class CreateTablesTable extends Migration
     public function up() {
         Schema::create('tables', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('table_group_id');
+            $table->unsignedBigInteger('table_group_id');
             $table->foreign('table_group_id')
                 ->references('id')->on('table_groups')
                 ->onDelete('cascade');
